+++ src/WinSmtpRelay.Storage/ServiceCollectionExtensions.cs
@@ -36,6 +36,7 @@ public static class ServiceCollectionExtensions
         services.AddScoped<IStatisticsRetentionSettingsService, StatisticsRetentionSettingsService>();
         services.AddScoped<IDataRetentionSettingsService, DataRetentionSettingsService>();
         services.AddScoped<IRetentionService, RetentionService>();
+        services.AddScoped<IAdminCertificateService, AdminCertificateService>();
         services.AddScoped<IReportingSettingsService, ReportingSettingsService>();
         services.AddScoped<IDnsSettingsService, DnsSettingsService>();
         services.AddScoped<IMessageFilterService, MessageFilterService>();